User model

+++ project-management-tool/src/main/java/com/javafreelancedeveloper/projectmanagementtool/service/ProjectService.java
@@ -1,13 +1,16 @@
 package com.javafreelancedeveloper.projectmanagementtool.service;
 
 import com.javafreelancedeveloper.projectmanagementtool.domain.Project;
+import com.javafreelancedeveloper.projectmanagementtool.domain.User;
 import com.javafreelancedeveloper.projectmanagementtool.dto.FieldValidationErrorResponseDTO;
 import com.javafreelancedeveloper.projectmanagementtool.dto.ProjectDTO;
 import com.javafreelancedeveloper.projectmanagementtool.dto.ProjectListDTO;
 import com.javafreelancedeveloper.projectmanagementtool.exception.ProjectNotFoundException;
+import com.javafreelancedeveloper.projectmanagementtool.exception.UserNotFoundException;
 import com.javafreelancedeveloper.projectmanagementtool.exception.ValidationException;
 import com.javafreelancedeveloper.projectmanagementtool.mapper.ProjectMapper;
 import com.javafreelancedeveloper.projectmanagementtool.repository.ProjectRepository;
+import com.javafreelancedeveloper.projectmanagementtool.repository.UserRepository;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DataIntegrityViolationException;
@@ -21,12 +24,17 @@ import java.util.List;
 @AllArgsConstructor
 public class ProjectService {
 
-    private final ProjectRepository projectRepository;
 
-    public ProjectDTO saveProject(ProjectDTO projectDTO) {
+    private final ProjectRepository projectRepository;
+    private final UserRepository userRepository;
+
+
+    public ProjectDTO saveProject(ProjectDTO projectDTO, String username) {
         Project project = ProjectMapper.map(projectDTO);
         try {
+            User user = userRepository.findByUsername(username).orElseThrow(() -> new UserNotFoundException("User not found for provided username."));
             project.setCode(project.getCode().toUpperCase());
+            project.setUser(user);
             Project savedProject = projectRepository.save(project);
             ProjectDTO savedProjectDTO = ProjectMapper.map(savedProject);
             return savedProjectDTO;
@@ -35,7 +43,8 @@ public class ProjectService {
         }
     }
 
-    public ProjectDTO updateProject(ProjectDTO projectDTO) {
+
+    public ProjectDTO updateProject(ProjectDTO projectDTO, String username) {
         if (projectDTO.getId() == null) {
             throw new ValidationException(new FieldValidationErrorResponseDTO("id", "Project Id is required for updates."));
         }
@@ -44,6 +53,7 @@ public class ProjectService {
         if (!project.getCode().equals(projectDTO.getCode())) {
             throw new ValidationException(new FieldValidationErrorResponseDTO("code", "Project Code is inconsistent with id. Invalid input."));
         }
+        validateAuthorisedUser(project, username);
         project.setName(projectDTO.getName());
         project.setDescription(projectDTO.getDescription());
         project.setStartDate(projectDTO.getStartDate());
@@ -53,31 +63,39 @@ public class ProjectService {
         return savedProjectDTO;
     }
 
-    public ProjectDTO findByProjectCode(String projectCode) {
+
+    public ProjectDTO findByProjectCode(String projectCode, String username) {
         Project project = projectRepository.findByCode(projectCode);
         if (project == null) {
             throw new ProjectNotFoundException("No project found with code [" + projectCode + "]");
         }
+        validateAuthorisedUser(project, username);
         return ProjectMapper.map(project);
 
     }
 
-    public ProjectListDTO listProjects() {
+    public ProjectListDTO listProjects(String username) {
         List<ProjectDTO> projectList = new ArrayList<>();
-        Iterable<Project> projects = projectRepository.findAll();
+        List<Project> projects = projectRepository.findByUser_Username(username);
         projects.forEach(p -> projectList.add(ProjectMapper.map(p)));
         return new ProjectListDTO(projectList);
 
     }
 
-    public void deleteProject(String projectCode) {
+    public void deleteProject(String projectCode, String username) {
         Project project = projectRepository.findByCode(projectCode.toUpperCase());
         if (project == null) {
             throw new ProjectNotFoundException("No project found with code [" + projectCode + "]");
         }
+        validateAuthorisedUser(project, username);
         projectRepository.delete(project);
     }
 
+    private void validateAuthorisedUser(Project project, String username) {
+        if (!project.getUser().getUsername().equals(username)) {
+            throw new ProjectNotFoundException("Requested project not found for user " + username + ".");
+        }
+    }
 
 
 }
